Genericize package
Grml's grub loopback.cfg interface has become a de facto standard for Linux
distributions.  Because of that, the Debian grml-rescueboot package is
usable with more ISOs than just Grml, so this change renames the package
to grub-loopback-iso, genericizes it, and provides backwards compatibility
with grml-rescueboot.  It does the following:

  * Renames everything to grub-loopback-iso.
  * Changes default ISOs directory to /boot/isos.
  * Adds a transitional grml-rescueboot package.
  * Replaces README with a generic explanation.
  * Removes previous stock /etc/grub.d/42_grml instances.
  * Checks for and uses the "legacy" /etc/default/grml-rescueboot if
    it exists.
  * Even if /etc/default/grml-rescueboot is not customized, "legacy"
    /boot/grml will be checked for as well.
  * Allows $ISO_LOCATION to contain multiple search directories, mostly as
    a by-product of backwards compatibility.

Note that debian/changelog is not updated, though this change expects the
next release to be numbered 0.4.6.  If that is not the case, be sure to
update the Replaces: and Breaks: in debian/control.

Signed-off-by: Ryan Finnie <ryan@finnie.org>